COMMERCE-1475 talend - component - Reorganize - move utils and common code to commons module and adopt classes
