+++ modules/etl/talend/talend-commons/src/main/java/com/liferay/talend/commons/log/DebugUtils.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.talend.utils;
+package com.liferay.talend.commons.log;
 
 import java.util.ArrayList;
 import java.util.Arrays;
