+++ modules/etl/talend/talend-commons/src/main/java/com/liferay/talend/commons/oas/OpenAPIType.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.talend.openapi;
+package com.liferay.talend.commons.oas;
 
 /**
  * @author Igor Beslic
