+++ modules/etl/talend/talend-commons/src/main/java/com/liferay/talend/commons/schema/SchemaUtils.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.talend.utils;
+package com.liferay.talend.commons.schema;
 
 import java.util.ArrayList;
 import java.util.HashSet;
