+++ modules/etl/talend/talend-definition/src/main/java/com/liferay/talend/tliferayoutput/TLiferayOutputProperties.java
@@ -15,13 +15,13 @@
 package com.liferay.talend.tliferayoutput;
 
 import com.liferay.talend.LiferayBaseComponentDefinition;
+import com.liferay.talend.commons.log.DebugUtils;
+import com.liferay.talend.commons.schema.SchemaUtils;
 import com.liferay.talend.connection.LiferayConnectionResourceBaseProperties;
-import com.liferay.talend.exception.ExceptionUtils;
+import com.liferay.talend.properties.ExceptionUtils;
 import com.liferay.talend.resource.LiferayResourceProperties;
 import com.liferay.talend.runtime.LiferaySourceOrSinkRuntime;
 import com.liferay.talend.runtime.ValidatedSoSSandboxRuntime;
-import com.liferay.talend.utils.DebugUtils;
-import com.liferay.talend.utils.SchemaUtils;
 
 import java.io.IOException;
