+++ modules/etl/talend/talend-runtime/src/main/java/com/liferay/talend/runtime/client/RESTClient.java
@@ -14,11 +14,11 @@
 
 package com.liferay.talend.runtime.client;
 
+import com.liferay.talend.commons.exception.MalformedURLException;
+import com.liferay.talend.commons.util.URIUtils;
 import com.liferay.talend.connection.LiferayConnectionProperties;
-import com.liferay.talend.exception.MalformedURLException;
 import com.liferay.talend.runtime.client.exception.ConnectionException;
 import com.liferay.talend.runtime.client.exception.OAuth2Exception;
-import com.liferay.talend.utils.URIUtils;
 
 import java.io.StringReader;
 import java.io.StringWriter;
